New folder/requirements.txt
streamlit>=1.32
openai>=1.3.5
python-dotenv>=1.0.0
supabase>=1.0.3
requests
Pillow
